embedding-server/pyproject.toml
[project]
name = "embedding-server"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "pillow>=10.1.0",
    "requests>=2.31.0",
    "hf-xet>=1.2.0"
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
